+++ Handlers/Products/Delete/DeleteProductHandler.cs
@@ -16,6 +16,14 @@ public class DeleteProductHandler : ICommandHandler<Guid, Guid>
     
     public Result<Guid, Error> Handle(Guid productId)
     {
+        var containers = _repository
+            .GetWarehouses()
+            .SelectMany(c => c.Containers);
+
+        if (containers.Any(pi => pi.Product.Id == productId))
+            return new Error("Ошибка при удалении товара",
+                "Товар нельзя удалить, так как он содержится в одном из контейнеров");
+
         var result = _repository.DeleteProduct(productId);
         
         return result;
